refactor: split index.vue logic into composables
