+++ app/pages/index.vue
@@ -1,64 +1,16 @@
 <script setup lang="ts">
-import type { LocationResponse } from '~/shared/types'
-import { toZonedTime } from 'date-fns-tz'
-import OpeningHours from 'opening_hours'
-
-function getOpeningHoursStatus(
-  expression: string | null | undefined,
-  timezone: string | null | undefined,
-  reference: Date = new Date(),
-): { isOpen: boolean, message: string, nextChange: Date | null, variant: 'open' | 'closing-soon' | 'closed' | 'unavailable' } {
-  if (!expression || !timezone)
-    return { isOpen: false, message: 'Hours unavailable', nextChange: null, variant: 'unavailable' }
-
-  try {
-    const localDate = toZonedTime(reference, timezone)
-    const schedule = new OpeningHours(expression.trim())
-
-    const isOpen = schedule.getState(localDate)
-    const nextChange = schedule.getNextChange(localDate) || null
-
-    // Visual indicator helps users plan visits better
-    let variant: 'open' | 'closing-soon' | 'closed' | 'unavailable' = isOpen ? 'open' : 'closed'
-    let message = isOpen ? 'Open now' : 'Closed'
-
-    if (isOpen && nextChange) {
-      const timeUntilClose = nextChange.getTime() - localDate.getTime()
-      const oneHour = 60 * 60 * 1000
-      if (timeUntilClose <= oneHour) {
-        variant = 'closing-soon'
-        message = 'Closing soon'
-      }
-    }
-
-    return { isOpen, message, nextChange, variant }
-  }
-  catch {
-    return { isOpen: false, message: 'Hours unavailable', nextChange: null, variant: 'unavailable' }
-  }
-}
-
-const searchQuery = ref('')
-const autocompleteResults = ref<LocationResponse[]>([])
-const showAutocomplete = ref(false)
-const selectedLocation = ref<LocationResponse | null>(null)
-
-const params = useUrlSearchParams('history')
-
-// URL params for shareable filter state
-const selectedCategories = computed<string[]>({
-  get: () => params.categories ? (params.categories as string).split(',') : [],
-  set: (val) => {
-    params.categories = val.length ? val.join(',') : null as any
-  },
-})
-
-const filters = computed<string[]>({
-  get: () => params.filters ? (params.filters as string).split(',') : [],
-  set: (val) => {
-    params.filters = val.length ? val.join(',') : null as any
-  },
-})
+const { selectedCategories, filters, removeCategory } = useSearchFilters()
+const {
+  searchQuery,
+  autocompleteResults,
+  showAutocomplete,
+  selectedLocation,
+  locations,
+  pending,
+  fetchAutocomplete,
+  handleSubmit,
+  selectLocation,
+} = useLocationSearch()
 
 const { locale, locales } = useI18n()
 const router = useRouter()
@@ -78,95 +30,10 @@ async function changeLocale(code: string) {
 }
 
 const { data: categories, refresh: refreshCategories } = useFetch('/api/categories', {
-  query: {
-    q: searchQuery,
-  },
+  query: { q: searchQuery },
   immediate: false,
   watch: false,
 })
-
-const { data: locations, pending, refresh: refreshLocations } = useFetch('/api/search', {
-  query: {
-    q: computed(() => searchQuery.value),
-    uuid: computed(() => selectedLocation.value?.uuid),
-    openNow: computed(() => filters.value.includes('open_now') ? 'true' : undefined),
-    categories: computed(() => selectedCategories.value.length ? selectedCategories.value.join(',') : undefined),
-  },
-  transform: (data) => {
-    return data?.map(location => ({
-      ...location,
-      hoursStatus: getOpeningHoursStatus(
-        location.openingHours,
-        location.timezone,
-      ),
-    })) ?? []
-  },
-  immediate: false,
-})
-
-const { execute: fetchAutocomplete } = useFetch('/api/search/autocomplete', {
-  query: {
-    q: searchQuery,
-  },
-  immediate: false,
-  onResponse({ response }) {
-    if (response.ok) {
-      autocompleteResults.value = response._data
-      showAutocomplete.value = true
-    }
-  },
-})
-
-// Speeds up form submission by caching embedding during typing
-async function precomputeEmbedding(query: string) {
-  if (query.length < 2)
-    return
-
-  // Non-blocking - autocomplete doesn't wait for this
-  $fetch('/api/search/embed', {
-    method: 'POST',
-    body: { q: query },
-  }).catch(() => {
-    // Silent fail - precomputing is optional optimization
-  })
-}
-
-// Debounce prevents excessive API calls while typing
-watch(searchQuery, useDebounceFn(async (newQuery) => {
-  if (newQuery.length < 2) {
-    showAutocomplete.value = false
-    autocompleteResults.value = []
-    return
-  }
-
-  await fetchAutocomplete()
-
-  // Background precompute for faster form submission
-  precomputeEmbedding(newQuery)
-}, 300))
-
-async function handleSubmit() {
-  if (!searchQuery.value || searchQuery.value.length < 2)
-    return
-
-  showAutocomplete.value = false
-
-  // User clicked their query instead of a location - semantic search
-  selectedLocation.value = null
-
-  await refreshLocations()
-}
-
-function selectLocation(location: LocationResponse | null) {
-  selectedLocation.value = location
-  showAutocomplete.value = false
-
-  if (location) {
-    searchQuery.value = location.name
-  }
-
-  refreshLocations()
-}
 
 const comboboxOpen = ref(false)
 
@@ -177,10 +44,6 @@ watch(selectedCategories, (newVal, oldVal) => {
     comboboxOpen.value = false
   }
 })
-
-function removeCategory(categoryId: string) {
-  selectedCategories.value = selectedCategories.value.filter(id => id !== categoryId)
-}
 
 function getCategoryById(categoryId: string) {
   return categories.value?.find(c => c.id === categoryId)
